+++ app/src/main/AndroidManifest.xml
@@ -175,7 +175,7 @@
             android:screenOrientation="portrait" />
 
         <activity
-            android:name="com.going.presentation.profile.trip.tripprofiletag.changetag.ChangeTagActivity"
+            android:name="com.going.presentation.profile.participant.profiletag.changetag.ChangeTagActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
