+++ presentation/src/main/java/com/going/presentation/profile/participant/ParticipantProfileActivity.kt
@@ -132,7 +132,6 @@ class ParticipantProfileActivity :
             }
         })
 
-
         setFragmentHeight()
     }
