--- presentation/src/main/java/com/going/presentation/profile/participant/ParticipantProfileTagFragment.kt
@@ -1,13 +1,0 @@
-package com.going.presentation.profile.participant
-
-import android.os.Bundle
-import com.going.presentation.R
-import com.going.presentation.databinding.FragmentTripProfileTagBinding
-import com.going.ui.base.BaseFragment
-
-class ParticipantProfileTagFragment :
-    BaseFragment<FragmentTripProfileTagBinding>(R.layout.fragment_trip_profile_tag) {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-}
