+++ presentation/src/main/java/com/going/presentation/profile/participant/ParticipantProfileViewPagerAdapter.kt
@@ -2,8 +2,11 @@ package com.going.presentation.profile.participant
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.going.presentation.profile.participant.profilecharacter.ParticipantProfileCharacterFragment
+import com.going.presentation.profile.participant.profiletag.ParticipantProfileTagFragment
 
-class ParticipantProfileViewPagerAdapter(activity: ParticipantProfileActivity) : FragmentStateAdapter(activity) {
+class ParticipantProfileViewPagerAdapter(activity: ParticipantProfileActivity) :
+    FragmentStateAdapter(activity) {
     override fun getItemCount(): Int = 2
 
     override fun createFragment(position: Int): Fragment =
